# Conflicts:
#	app/src/main/java/com/delighted2wins/souqelkhorda/navigation/NavigationRoot.kt
#	app/src/main/res/values-ar/strings.xml
#	app/src/main/res/values/strings.xml

+++ app/src/main/java/com/delighted2wins/souqelkhorda/features/myorders/presentation/screen/OrdersScreen.kt
@@ -1,5 +1,6 @@
 package com.delighted2wins.souqelkhorda.features.myorders.presentation.screen
 
+import android.util.Log
 import androidx.compose.foundation.ExperimentalFoundationApi
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Arrangement
@@ -34,13 +35,13 @@ import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import androidx.compose.ui.platform.LocalConfiguration
 import androidx.compose.ui.platform.LocalLayoutDirection
+import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.LayoutDirection
 import androidx.hilt.lifecycle.viewmodel.compose.hiltViewModel
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
+import com.delighted2wins.souqelkhorda.R
 import com.delighted2wins.souqelkhorda.core.enums.OrderSource
-import com.delighted2wins.souqelkhorda.features.market.presentation.contract.MarketEffect
-import com.delighted2wins.souqelkhorda.features.market.presentation.contract.MarketIntent
-import com.delighted2wins.souqelkhorda.features.myorders.presentation.component.DeleteConfirmationDialog
+import com.delighted2wins.souqelkhorda.core.components.ConfirmationDialog
 import com.delighted2wins.souqelkhorda.features.myorders.presentation.contract.MyOrdersEffect
 import com.delighted2wins.souqelkhorda.features.myorders.presentation.contract.MyOrdersIntents
 import com.delighted2wins.souqelkhorda.features.myorders.presentation.viewmodel.MyOrdersViewModel
@@ -52,7 +53,7 @@ fun OrdersScreen(
     innerPadding: PaddingValues = PaddingValues(),
     snackBarHostState: SnackbarHostState,
     viewModel: MyOrdersViewModel = hiltViewModel(),
-    onDetailsClick: (String, String) -> Unit = { _, _ -> }
+    onDetailsClick: (orderId: String, ownerId: String, buyerId: String, source: OrderSource) -> Unit,
 ) {
     val state by viewModel.state.collectAsStateWithLifecycle()
     val systemIsRtl = LocalConfiguration.current.layoutDirection == LayoutDirection.Rtl.ordinal
@@ -67,6 +68,7 @@ fun OrdersScreen(
     val isRtl: Boolean = LocalLayoutDirection.current == LayoutDirection.Rtl
     var showDeleteDialog by remember { mutableStateOf(false) }
     var orderToDelete by remember { mutableStateOf<String?>(null) }
+    var selectedChip by remember { mutableStateOf("Sells") }
 
 
     LaunchedEffect(Unit) {
@@ -160,7 +162,9 @@ fun OrdersScreen(
                     state.saleOrders,
                     state.isLoading,
                     state.error,
-                    onDetailsClick,
+                    onDetailsClick = { orderId, ownerId ->
+                        onDetailsClick(orderId, ownerId, "", OrderSource.COMPANY)
+                    },
                     onDeclineClick = { orderId ->
                         orderToDelete = orderId
                         showDeleteDialog = true
@@ -170,19 +174,32 @@ fun OrdersScreen(
                 1 -> MarketOrdersScreen(
                     state = state,
                     onChipSelected = { chip ->
+                        selectedChip = chip
                         when (chip) {
                             "Sells" -> viewModel.onIntent(MyOrdersIntents.LoadSells)
                             "Offers" -> viewModel.onIntent(MyOrdersIntents.LoadOffers)
                         }
                     },
-                    onDetailsClick,
+                    onDetailsClick = { orderId, ownerId ->
+                        val buyerId = state.currentBuyerId
+                        val source = when (selectedChip) {
+                            "Sells" -> OrderSource.SALES
+                            "Offers" -> OrderSource.OFFERS
+                            else -> OrderSource.MARKET
+                        }
+                        Log.d("OrdersScreen", "OrdersScreen: $source")
+                        onDetailsClick(orderId, ownerId, buyerId.orEmpty(), source)
+                    },
                     systemIsRtl
                 )
             }
         }
         if (showDeleteDialog) {
-            DeleteConfirmationDialog(
-                isRtl = isRtl,
+            ConfirmationDialog(
+//                isRtl = isRtl,
+                title = stringResource(R.string.confirm_delete),
+                message = stringResource(R.string.are_you_sure_you_want_to_delete_this_order),
+                confirmLabel = stringResource(R.string.delete),
                 onConfirm = {
                     orderToDelete?.let { viewModel.onIntent(MyOrdersIntents.DeleteCompanyOrder(it)) }
                     showDeleteDialog = false
